Started treating models as a submodule.
Also, added a warning sign to any users attempting to use this branch.
The reason we want models out of the repo is because they add a lot of
overhead when cloning with Git. Instead, by having it as a submodule,
the user can freely choose if he/she wants to pull down the models or
not. Since this might be used in headless mode, this is not always a
desire. We might have to clean the Git history to truly fix things...