Generated mips for sRGB textures as well.
We do this by interpreting the textures as RGBA initially, and then, in
the shader, do the fast sRGB approximation. This makes sure the sampler
information is compliant with what's in the glTF, and actually generate
mips when it should. It costs a bit extra (GLSL pow) but it's worth it.

I've compared both of them, and there is very little qualitative diffs.